pyproject.toml
[tool.poetry]
name = "janjic-kurs-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "janjic_kurs_7"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
stripe = "^7.5.0"
celery = "^5.3.5"
eventlet = "^0.33.3"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
telegram = "^0.0.1"
telebot = "^0.0.5"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
